+++ Elit3d/src/Tools/System/Profiler.cpp
@@ -82,7 +82,7 @@ void ProfilerTimer::Stop()
 	long long start = std::chrono::time_point_cast<std::chrono::microseconds>(start_timepoint).time_since_epoch().count();
 	long long end = std::chrono::time_point_cast<std::chrono::microseconds>(end_timepoint).time_since_epoch().count();
 
-	uint32_t threadID = std::hash<std::thread::id>{}(std::this_thread::get_id());
+	uint32_t threadID = (uint32_t)std::hash<std::thread::id>{}(std::this_thread::get_id());
 	Profiler::Get().Insert({ name, start, end, threadID });
 
 	if (log)
